Move VolumeCornerIterator to DomainHelpers
